More reuse of New Branch page
