+++ GitAutomation/package.json
@@ -6,12 +6,9 @@
   "scripts": {
     "start": "webpack --watch",
     "build": "webpack",
-    "gql-generation":
-      "gql2ts generated/graphql.json -o web-scripts/graphql.generated.d.ts -n GitAutomationGQL",
-    "gql:nix":
-      "(mkdir generated || echo 1) && cd ../GitAutomation.GraphQL && (dotnet build || echo 1) && dotnet run --no-build > ../GitAutomation/generated/graphql.json && cd ../GitAutomation && generated/graphql.json | yarn gql-generation",
-    "gql:win":
-      "(mkdir generated || echo 1) && cd ../GitAutomation.GraphQL && (dotnet build || echo 1) && dotnet run --no-build > ../GitAutomation/generated/graphql.json && cd ../GitAutomation && yarn gql-generation < generated/graphql.json",
+    "gql-generation": "gql2ts generated/graphql.json -o web-scripts/graphql.generated.d.ts -n GitAutomationGQL",
+    "gql:nix": "(mkdir generated || echo 1) && cd ../GitAutomation.GraphQL && (dotnet build || echo 1) && dotnet run --no-build > ../GitAutomation/generated/graphql.json && cd ../GitAutomation && generated/graphql.json | yarn gql-generation",
+    "gql:win": "(mkdir generated || echo 1) && cd ../GitAutomation.GraphQL && (dotnet build || echo 1) && dotnet run --no-build > ../GitAutomation/generated/graphql.json && cd ../GitAutomation && yarn gql-generation < generated/graphql.json",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Matt DeKrey",
@@ -43,6 +40,7 @@
     "html-loader": "0.5.1",
     "html-minifier": "^3.5.5",
     "html-minifier-loader": "^1.4.0",
+    "immer": "1.1.0",
     "ramda": "^0.24.1",
     "raw-loader": "^0.5.1",
     "react": "16.2.0",
