--- GitAutomation/web-scripts/new-branch/bind-save-button.ts
@@ -1,30 +1,0 @@
-import { Observable } from "../utils/rxjs";
-
-import { createBranch } from "../api/basics";
-
-export interface ISaveData {
-  downstream: string[];
-  upstream: string[];
-  branchName: string;
-  upstreamMergePolicy: GitAutomationGQL.IUpstreamMergePolicyEnum;
-  branchType: GitAutomationGQL.IBranchGroupTypeEnum;
-}
-
-export const doSave = (data: Observable<ISaveData>) => {
-  // save
-  return data
-    .take(1)
-    .map(({ branchName, upstream, upstreamMergePolicy, branchType }) => {
-      return {
-        branchName,
-        requestBody: {
-          upstreamMergePolicy,
-          branchType,
-          addUpstream: upstream
-        }
-      };
-    })
-    .switchMap(({ branchName, requestBody }) =>
-      createBranch(branchName, requestBody).map(_ => branchName)
-    );
-};
