--- GitAutomation/web-scripts/new-branch/new-branch-check-listing.row.html
@@ -1,7 +1,0 @@
-<td data-locator="branch" data-class="branchName" />
-<!-- <td data-locator="downstream-branches" data-class="checkboxCell">
-  <input type="checkbox" data-locator="check" />
-</td> -->
-<td data-locator="upstream-branches" data-class="checkboxCell">
-  <input type="checkbox" data-locator="check" />
-</td>
