--- GitAutomation/web-scripts/new-branch/new-branch.layout.html
@@ -1,61 +1,0 @@
-<h1>New Branch</h1>
-<h3>Settings</h3>
-<section>
-    <section>
-        <label>
-            Branch Name
-            <input type="text" data-locator="branch-name" />
-        </label>
-    </section>
-    <section data-class="fieldSection">
-        <label>
-            Branch Type
-            <select data-locator="branch-type">
-                <option value="Feature">Feature</option>
-                <option value="ReleaseCandidate">Release Candidate</option>
-                <option value="ServiceLine">Service Line</option>
-                <option value="Infrastructure">Infrastructure</option>
-                <option value="Integration">Integration</option>
-                <option value="Hotfix">Hotfix</option>
-            </select>
-        </label>
-    </section>
-    <section data-class="fieldSection">
-        <label>
-            Upstream Policy
-            <select data-locator="upstream-merge-policy">
-                <option value="None">None (merge normally)</option>
-                <option value="MergeNextIteration">Create new branch Iteration</option>
-                <option value="ForceFresh">Force update base branch</option>
-            </select>
-        </label>
-        <p data-class="hint">Used only with "Release Candidates"; will make new branches that contain only upstream commits</p>
-    </section>
-</section>
-<h3>Other Branches</h3>
-<table data-class="otherBranchTable">
-    <thead>
-        <tr>
-            <td></td>
-            <!-- <th data-class="rotateHeader">
-                <div>
-                    <span>Downstream</span>
-                </div>
-            </th> -->
-            <th data-class="rotateHeader">
-                <div>
-                    <span>Upstream</span>
-                </div>
-            </th>
-        </tr>
-    </thead>
-    <tbody data-locator="other-branches">
-    </tbody>
-</table>
-
-<section>
-    <h5>Preview graph</h5>
-    <svg data-locator="hierarchy-container-preview" width="800" height="100" style="max-height: 70vh;"></svg>
-</section>
-<button type="button" data-locator="home">Cancel</button>
-<button type="button" data-locator="save" data-role="create administrate">Save</button>
